changed Item class to ItemCore and created new wrapper Item class
Changed the name of the Item class to ItemCore.
This allows me to add an inventory to the item class and avoid a circular dependency. Now Inventory can be added to clases such as Item, Room, etc so they can have an inventory of items in them. Each class will get a wrapper such as item which manages the interface, while the Inventory will be managed by the inventory class.  This will make it easy to add functions to the Inventory class that swap items between them without needing to make different kinds of inventories for different kinds of items (ie: rooms, player inventory, items that contain other items, etc).